Added ability to delete record and check access to admin commands. Removed "id" key from json file.

+++ app/handlers.py
@@ -6,6 +6,7 @@ from aiogram.types import Message, CallbackQuery, ForceReply
 from aiogram.filters import Command
 from aiogram.fsm.context import FSMContext
 
+import config
 from main import bot
 from app import keyboards, json_file
 from app.states import ShameData
@@ -13,9 +14,18 @@ from app.states import ShameData
 router = Router()
 
 
+async def is_user_has_insufficient_rights(callback) -> bool:
+    if callback.from_user.id != config.admin:
+        await callback.answer('У вас недостаточно прав')
+        return True
+    else:
+        return False
+
+
 @router.message(Command('start', 'shame'))
 async def bot_call_handler(message: Message, state: FSMContext):
     await state.update_data(message_id=message.message_id)
+    await state.update_data(username=message.from_user.username)
     await message.reply(text='Выберите группу:',
                         reply_markup=keyboards.group_selection)
 
@@ -31,9 +41,9 @@ async def home_handler(callback: CallbackQuery, state: FSMContext):
 
 @router.callback_query(F.data == 'delete_message')
 async def delete_message_handler(callback: CallbackQuery, state: FSMContext):
-    state_data = await state.get_state()
-    if state_data is not None:
-        await state.set_state(state=None)
+    if await is_user_has_insufficient_rights(callback):
+        return
+    await state.clear()
     await callback.message.delete()
 
 
@@ -55,17 +65,17 @@ async def check_history_handler(callback: CallbackQuery, state: FSMContext):
             data = await state.get_data()
             shame_data = json_file.get_the_shame_data('shame.json', data['group_name'], data['shame_id'])
             await callback.message.edit_text(text=f'Выбрана группа: {data["group_name"]}\n'
-                                                  f'Выбрано действие: история\n\n'
+                                                  f'Выбрано действие: Просмотр истории\n\n'
                                                   f'{shame_data}',
                                              reply_markup=keyboards.check_history)
         else:
             await callback.message.edit_text(text=f'Выбрана группа: {data["group_name"]}\n'
-                                                  f'Выбрано действие: история\n'
+                                                  f'Выбрано действие: Просмотр истории\n'
                                                   f'Вы ещё не записали ни одного позора',
                                              reply_markup=keyboards.go_home)
     else:
         await callback.message.edit_text(text=f'Выбрана группа: {data["group_name"]}\n'
-                                              f'Выбрано действие: история\n'
+                                              f'Выбрано действие: Просмотр истории\n'
                                               f'Вы ещё не записали ни одного позора',
                                          reply_markup=keyboards.go_home)
 
@@ -81,7 +91,7 @@ async def back_history_handler(callback: CallbackQuery, state: FSMContext):
         data = await state.update_data(shame_id=next_id)
     shame_data = json_file.get_the_shame_data('shame.json', data['group_name'], data['shame_id'])
     await callback.message.edit_text(text=f'Выбрана группа: {data["group_name"]}\n'
-                                          f'Выбрано действие: история\n\n'
+                                          f'Выбрано действие: Просмотр истории\n\n'
                                           f'{shame_data}',
                                      reply_markup=keyboards.check_history)
 
@@ -98,9 +108,36 @@ async def forward_history_handler(callback: CallbackQuery, state: FSMContext):
         data = await state.update_data(shame_id=next_id)
     shame_data = json_file.get_the_shame_data('shame.json', data['group_name'], data['shame_id'])
     await callback.message.edit_text(text=f'Выбрана группа: {data["group_name"]}\n'
-                                          f'Выбрано действие: история\n\n'
+                                          f'Выбрано действие: Просмотр истории\n\n'
                                           f'{shame_data}',
                                      reply_markup=keyboards.check_history)
+
+
+@router.callback_query(F.data == 'delete_record')
+async def delete_record_handler(callback: CallbackQuery, state: FSMContext):
+    if await is_user_has_insufficient_rights(callback):
+        return
+    data = await state.get_data()
+    shame_data = json_file.get_the_shame_data('shame.json', data['group_name'], data['shame_id'])
+    await callback.message.edit_text(text=f'Вы уверены, что хотите удалить запись:\n\n'
+                                          f'{shame_data}',
+                                     reply_markup=keyboards.action_check)
+
+
+@router.callback_query(F.data == 'positive_check')
+async def positive_check_handler(callback: CallbackQuery, state: FSMContext):
+    if await is_user_has_insufficient_rights(callback):
+        return
+    data = await state.get_data()
+    json_file.delete_the_record('shame.json', data['group_name'], data['shame_id'])
+    await check_history_handler(callback, state)
+
+
+@router.callback_query(F.data == 'negative_check')
+async def positive_check_handler(callback: CallbackQuery, state: FSMContext):
+    if await is_user_has_insufficient_rights(callback):
+        return
+    await check_history_handler(callback, state)
 
 
 @router.callback_query(F.data == 'add_shame')
@@ -108,7 +145,7 @@ async def add_shame_handler(callback: CallbackQuery, state: FSMContext):
     data = await state.get_data()
     await bot.send_message(chat_id=callback.message.chat.id,
                            text=f'Выбрана группа: {data["group_name"]}\n'
-                                f'Выбрано действие: +позор\n'
+                                f'Выбрано действие: Добавить позор\n'
                                 f'Опишите ситуацию',
                            reply_to_message_id=data['message_id'],
                            reply_markup=ForceReply(selective=True)
@@ -119,11 +156,14 @@ async def add_shame_handler(callback: CallbackQuery, state: FSMContext):
 
 @router.message(ShameData.description)
 async def description_handler(message: Message, state: FSMContext):
-    if message.reply_to_message is not None:
+    data = await state.get_data()
+    if (message.reply_to_message is not None) and (message.from_user.id == message.reply_to_message.chat.id):
         await state.update_data(description=message.text)
         delta = datetime.timedelta(hours=3, minutes=0)
         data = await state.update_data(message_datetime=(message.date + delta).strftime('%d.%m.%y %H:%M'))
         json_file.add_record_to_json('shame.json', data['group_name'], data['message_datetime'], data['description'])
         count = json_file.get_the_record_number('shame.json', data['group_name'])
         await message.reply(text=f'Для {data["group_name"]} добавлен позор №{count} от {data["message_datetime"]}\n\n')
-        await state.set_state(state=None)
+        await state.clear()
+    else:
+        await message.reply(f'У вас недостаточно прав. Ожидание текста от @{data["username"]}')
